requirements.txt
appdirs
bidict
dill
docopt
numpy
pandas
ply
pykwalify
python-dateutil
pytz
PyYAML
scipy
six
jinja2
tqdm
requests
markdown
